requirements.txt
PIL==1.1.7
argparse==1.2.1
distribute==0.6.34
google-api-python-client==1.1
httplib2==0.8
ipython==0.13.2
numpy==1.8.0.dev-Unknown
pygame==1.9.2pre
python-gflags==2.0
wsgiref==0.1.2
